refined the spaces output
