+++ src/whiteboards-info/model/whiteboardSpaceInfo.ts
@@ -1,0 +1,11 @@
+export class WhiteboardSpaceInfo {
+  spaceNameID = ''; // nameID of space
+  spaceID = ''; // UUID
+  SpaceLevel = ''; // "L0", "L1", "L2"
+  SpaceVisibility = ''; // "DEMO", "ACTIVE", "ARCHIVED"
+  CalloutName = ''; // nameID of the callout
+  CalloutID = ''; // UUID
+  WhiteboardName = ''; // Display name of the whiteboard
+  WhiteboardID = ''; // UUID
+  WhiteboardURL = ''; // The profile.url from whiteboard
+}
