--- src/whiteboards-info/whiteboards-info.graphql
@@ -1,195 +1,0 @@
-query whiteboardsInfo {
-  platformAdmin {
-    spaces(filter: { visibilities: [DEMO] }) {
-      id
-      nameID
-      visibility
-      about {
-        profile {
-          displayName
-        }
-      }
-      collaboration {
-        id
-        calloutsSet {
-          id
-          callouts {
-            id
-            nameID
-            framing {
-              id
-              whiteboard {
-                id
-                nameID
-                profile {
-                  id
-                  displayName
-                  url
-                }
-              }
-            }
-          }
-        }
-      }
-      subspaces {
-        id
-        nameID
-        collaboration {
-          id
-          calloutsSet {
-            id
-            callouts {
-              id
-              framing {
-                id
-                whiteboard {
-                  id
-                  profile {
-                    id
-                    url
-                  }
-                }
-              }
-            }
-          }
-        }
-        subspaces {
-          id
-          nameID
-          collaboration {
-            id
-            calloutsSet {
-              id
-              callouts {
-                id
-                framing {
-                  id
-                  whiteboard {
-                    id
-                    profile {
-                      id
-                      url
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-
-  platformAdmin {
-    innovationPacks {
-      id
-      nameID
-      provider {
-        id
-        profile {
-          displayName
-        }
-      }
-      templatesSet {
-        templates {
-        id
-        type
-        profile {
-          displayName
-        }
-        callout {
-          id
-          nameID
-          framing {
-            id
-            whiteboard {
-              id
-              nameID
-              profile {
-                id
-                displayName
-                url
-              }
-            }
-          }
-        }
-        contentSpace {
-          id
-          collaboration {
-            id
-            calloutsSet {
-              id
-              callouts {
-                id
-                nameID
-                framing {
-                  id
-                  whiteboard {
-                    id
-                    nameID
-                    profile {
-                      id
-                      displayName
-                      url
-                    }
-                  }
-                }
-              }
-            }
-          }
-          subspaces {
-            id
-            collaboration {
-              id
-              calloutsSet {
-                id
-                callouts {
-                  id
-                  nameID
-                  framing {
-                    id
-                    whiteboard {
-                      id
-                      nameID
-                      profile {
-                        id
-                        displayName
-                        url
-                      }
-                    }
-                  }
-                }
-              }
-            }
-            subspaces {
-              id
-              collaboration {
-                id
-                calloutsSet {
-                  id
-                  callouts {
-                    id
-                    nameID
-                    framing {
-                      id
-                      whiteboard {
-                        id
-                        nameID
-                        profile {
-                          id
-                          displayName
-                          url
-                        }
-                      }
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-    }
-  }
-}
